+++ src/main/java/com/epam/aidial/metric/service/AbstractQueryBuilder.java
@@ -28,7 +28,7 @@ public abstract class AbstractQueryBuilder<C, T extends TableDeclaration> {
     protected final Map<String, T> tableDeclarations;
     protected final AbstractDatasetConfiguration datasetConfiguration;
     protected final Map<Expression, String> expressionToOuterColumnNames = new HashMap<>();
-    protected final Map<String, Expression> aliasToExpression = new HashMap<>();
+    protected final Map<String, Expression> nameToExpression = new HashMap<>();
     protected final TemporalNameGenerator temporalNameGenerator;
 
     @SuppressWarnings("unchecked")
@@ -45,7 +45,6 @@ public abstract class AbstractQueryBuilder<C, T extends TableDeclaration> {
         if (completable instanceof Query query) {
 
             resolveExpressionsToOuterColumnNames(query.getExpressions());
-            resolveAliases(query.getExpressions());
 
             if (isWindowColumnAggregationQuery(query)) {
                 return buildWindowColumnAggregationQuery(query);
@@ -99,7 +98,7 @@ public abstract class AbstractQueryBuilder<C, T extends TableDeclaration> {
             return true;
         }
         if (expression instanceof Column column) {
-            var originalExpression = aliasToExpression.get(column.getName());
+            var originalExpression = nameToExpression.get(column.getName());
             return originalExpression instanceof GroupFunctionCall;
         }
         return false;
@@ -143,7 +142,7 @@ public abstract class AbstractQueryBuilder<C, T extends TableDeclaration> {
         if (expression instanceof Alias alias) {
             return alias.getExpression();
         } else if (expression instanceof Column column) {
-            var aliasedExpression = aliasToExpression.get(column.getName());
+            var aliasedExpression = nameToExpression.get(column.getName());
             if (aliasedExpression instanceof GroupFunctionCall groupFunctionCall) {
                 return groupFunctionCall;
             }
@@ -230,13 +229,15 @@ public abstract class AbstractQueryBuilder<C, T extends TableDeclaration> {
             if (expression instanceof Alias alias) {
                 expressionToOuterColumnNames.put(alias, alias.getName());
                 expressionToOuterColumnNames.put(alias.getExpression(), alias.getName());
+                nameToExpression.put(alias.getName(), alias.getExpression());
             } else if (expression instanceof Column column) {
-                var aliasedExpression = aliasToExpression.get(column.getName());
+                var aliasedExpression = nameToExpression.get(column.getName());
                 if (aliasedExpression instanceof GroupFunctionCall) {
                     expressionToOuterColumnNames.put(expression, getNewTemporaryName(TEMPORAL_COLUMN_NAME));
                 } else {
                     expressionToOuterColumnNames.put(expression, column.getName());
                 }
+                nameToExpression.put(column.getName(), expression);
             } else if (expression instanceof AggregationFunctionCall) {
                 expressionToOuterColumnNames.put(expression, getNewTemporaryName(TEMPORAL_COLUMN_NAME));
             } else if (expression instanceof GroupFunctionCall) {
@@ -247,12 +248,22 @@ public abstract class AbstractQueryBuilder<C, T extends TableDeclaration> {
         }
     }
 
-    protected void resolveAliases(List<Expression> expressions) {
-        for (Expression expression : expressions) {
-            if (expression instanceof Alias alias) {
-                aliasToExpression.put(alias.getName(), alias.getExpression());
+    /**
+     * Resolves a sort/orderBy expression to its outer column name.
+     * First tries direct object lookup, then resolves by name through nameToExpression.
+     */
+    protected String resolveOrderByColumnName(Expression expression) {
+        var columnName = expressionToOuterColumnNames.get(expression);
+        if (columnName != null) {
+            return columnName;
+        }
+        if (expression instanceof Column column) {
+            var resolved = nameToExpression.get(column.getName());
+            if (resolved != null) {
+                return expressionToOuterColumnNames.get(resolved);
             }
         }
+        return null;
     }
 
     protected String getNewTemporaryName(String prefix) {
